add ProtoV6ProviderFactories to TestCase
In order to support protocol v6 providers in the test suite, we add another
field to TestCase along the same lines as ProtoV5ProviderFactories. This is
plumbed through as yet another argument in runProviderCommand. We hope that by
the time another protocol major version is released, a new, more flexible test
framework will be available.

plugin.ServeOpts and plugin.DebugServe also expected all provider servers to be
v5 servers, so we also add the GRPCProviderV6Func field to ServeOpts to keep
it backwards compatible. If both GRPCProviderFunc and GRPCProviderV6Func are
set, the v6 server will be ignored.